Make the game mode a part of the game options
The game mode is now written to the database at game creation.
In future changes the game mode will be modifiable.
When the game starts, the Lobby view checks which game mode
we're in and redirects to the right view.
RendezvousGame is currently a dummy view, but will house the
main game logic for Rendezvous (analogous to Game.vue, which
is not called ClassicGame.vue).